rename grpc pkg to endpoint and use more descriptive file names.
